Fix issues with note attack: 1) ensure we don't waste time slewing from the previous note played on the same pad (reset fxdPrevPressure on new note and also slew full-speed during first 50ms); 2) reset noteInitialMaxValueZHi when new note starts; 3) re-arrange order to send note-off first and set previousValueZHi  before calling preSendLoudness(); 4) always use velocity-to-pressure so that we have a better estimate of initial pressure values

+++ linnstrument-firmware.ino
@@ -372,8 +372,8 @@ int :5;
   byte velocityZ:7;                          // the Z value with velocity sensitivity
 
   byte noteInitialVelocity:8;
-  byte noteInitialMaxValueZHi:10;
-  byte previousValueZHi:10;
+  unsigned short noteInitialMaxValueZHi:10;
+  unsigned short previousValueZHi:10;
 };
 TouchInfo touchInfo[MAXCOLS][MAXROWS];       // store as much touch information instances as there are cells
